add UnlockedManifest class
